dash/requirements.txt
dash
dash_core_components
dash_html_components
mysql-connector-python
plotly.express
pandas